terhal_server/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
astroid==3.0.1; python_full_version >= '3.8.0'
attrs==23.1.0; python_version >= '3.7'
cachecontrol==0.13.1; python_version >= '3.7'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
colorama==0.4.6; sys_platform == 'win32'
cryptography==41.0.5
dill==0.3.7; python_version >= '3.11'
django==4.2.7; python_version >= '3.8'
django-filter==23.3; python_version >= '3.7'
django-json-widget==1.1.1
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-csv==3.0.1
drf-spectacular==0.26.5; python_version >= '3.6'
firebase-admin==6.2.0; python_version >= '3.7'
future==0.18.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core[grpc]==2.14.0; platform_python_implementation != 'PyPy'
google-api-python-client==2.106.0; python_version >= '3.7'
google-auth==2.23.4; python_version >= '3.7'
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3; python_version >= '3.7'
google-cloud-firestore==2.13.1; platform_python_implementation != 'PyPy'
google-cloud-storage==2.13.0; python_version >= '3.7'
google-crc32c==1.5.0; python_version >= '3.7'
google-resumable-media==2.6.0; python_version >= '3.7'
googleapis-common-protos==1.61.0; python_version >= '3.7'
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==21.2.0; python_version >= '3.5'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
isort==5.12.0; python_full_version >= '3.8.0'
jsonschema==4.19.2; python_version >= '3.8'
jsonschema-specifications==2023.7.1; python_version >= '3.8'
mccabe==0.7.0; python_version >= '3.6'
msgpack==1.0.7; python_version >= '3.8'
numpy==1.26.1; python_version < '3.13' and python_version >= '3.9'
packaging==23.2; python_version >= '3.7'
pandas==2.1.2; python_version >= '3.9'
pillow==10.1.0; python_version >= '3.8'
platformdirs==3.11.0; python_version >= '3.7'
proto-plus==1.22.3; python_version >= '3.6'
protobuf==4.25.0; python_version >= '3.7'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyjwt[crypto]==2.8.0; python_version >= '3.7'
pylint==3.0.2; python_full_version >= '3.8.0'
pylint-django==2.5.5; python_version >= '3.7' and python_version < '4.0'
pylint-plugin-utils==0.8.2; python_version >= '3.7' and python_version < '4.0'
pyparsing==3.1.1; python_version >= '3.1'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.30.2; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rpds-py==0.12.0; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
tomlkit==0.12.2; python_version >= '3.7'
tzdata==2023.3; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.7; python_version >= '3.7'
